requirements.txt
Flask
PyQt5
pandas
pyngrok
python-dotenv